refactor: move resolver code to new dir
