+++ src/schema/v2/homeView/resolvers/activityResolvers.ts
@@ -1,6 +1,6 @@
 import type { GraphQLFieldResolver } from "graphql"
 import type { ResolverContext } from "types/graphql"
-import { NotificationsConnection } from "../notifications"
+import { NotificationsConnection } from "../../notifications"
 
 /*
  * Resolvers for home view activity sections
