+++ src/schema/v2/homeView/resolvers/featuredFairsResolver.ts
@@ -1,15 +1,13 @@
 import { GraphQLFieldResolver } from "graphql"
 import { ResolverContext } from "types/graphql"
-import { HomePageSalesModuleType } from "../home/home_page_sales_module"
+import { HomePageFairsModuleType } from "../../home/home_page_fairs_module"
 import { connectionFromArray } from "graphql-relay"
 
-export const SalesResolver: GraphQLFieldResolver<any, ResolverContext> = async (
-  parent,
-  args,
-  context,
-  info
-) => {
-  const { results: resolver } = HomePageSalesModuleType.getFields()
+export const FeaturedFairsResolver: GraphQLFieldResolver<
+  any,
+  ResolverContext
+> = async (parent, args, context, info) => {
+  const { results: resolver } = HomePageFairsModuleType.getFields()
 
   if (!resolver?.resolve) {
     return []
